@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:projeto_api_geo/Controller/weather_controller.dart';
 import 'package:projeto_api_geo/Service/city_db_service.dart';
+
 import '../Model/city_model.dart';
 import 'details_weather_screen.dart';
 
 class SearchScreen extends StatefulWidget {
-  const SearchScreen({Key? key}) : super(key: key);
+  const SearchScreen({Key? key});
 
   @override
   State<SearchScreen> createState() => _SearchScreenState();
@@ -20,9 +21,7 @@ class _SearchScreenState extends State<SearchScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text("Pesquisa Por Cidade"),
-      ),
+      appBar: AppBar(title: const Text("Pesquisa Por Cidade")),
       body: Padding(
         padding: const EdgeInsets.all(12),
         child: Column(
@@ -42,16 +41,14 @@ class _SearchScreenState extends State<SearchScreen> {
                         return null;
                       },
                     ),
-                    const SizedBox(
-                      height: 20,
-                    ),
+                    const SizedBox(height: 20),
                     ElevatedButton(
-                      onPressed: () async {
+                      onPressed: () {
                         if (_formKey.currentState!.validate()) {
-                          await _findCity(_cityController.text);
+                          _findCity(_cityController.text);
                         }
                       },
-                      child: const Text("Search"),
+                      child: const Text("Pesquisar"),
                     ),
                     const SizedBox(height: 20),
                   ],
@@ -60,32 +57,30 @@ class _SearchScreenState extends State<SearchScreen> {
             ),
             Expanded(
               child: FutureBuilder<List<City>>(
-  future: _dbService.getAllCities().then((list) => list.map((e) => City.fromMap(e)).toList()),
-  builder: (context, AsyncSnapshot<List<City>> snapshot) {
-    if (snapshot.connectionState == ConnectionState.waiting) {
-      return const Center(child: CircularProgressIndicator());
-    } else if (snapshot.hasError) {
-      return Center(child: Text('Erro: ${snapshot.error}'));
-    } else if (!snapshot.hasData || snapshot.data!.isEmpty) {
-      return const Center(child: Text('Lista Vazia'));
-    } else {
-      List<City> cities = snapshot.data!;
-      return ListView.builder(
-        itemCount: cities.length,
-        itemBuilder: (context, index) {
-          final city = cities[index];
-          return ListTile(
-            title: Text(city.cityName),
-            onTap: () {
-              _findCity(city.cityName); // Chama _findCity com o nome da cidade
-            },
-          );
-        },
-      );
-    }
-  },
-)
-
+                future: _dbService.getAllCities(),
+                builder: (context, snapshot) {
+                  if (snapshot.connectionState == ConnectionState.waiting) {
+                    return const Center(child: CircularProgressIndicator());
+                  } else if (snapshot.hasError) {
+                    return const Center(child: Text("Erro ao carregar histórico"));
+                  } else if (!snapshot.hasData || snapshot.data!.isEmpty) {
+                    return const Center(child: Text("Sem Histórico"));
+                  } else {
+                    return ListView.builder(
+                      itemCount: snapshot.data!.length,
+                      itemBuilder: (context, index) {
+                        final city = snapshot.data![index];
+                        return ListTile(
+                          title: Text(city.cityName),
+                          onTap: () {
+                            _findCity(city.cityName);
+                          },
+                        );
+                      },
+                    );
+                  }
+                },
+              ),
             ),
           ],
         ),
@@ -93,20 +88,32 @@ class _SearchScreenState extends State<SearchScreen> {
     );
   }
 
-Future<void> _findCity(String cityName) async {
-  if (await _controller.findCity(cityName)) {
-    City city = City(cityName: cityName, favoriteCities: 0); // Cria a instância de City
-    await _dbService.insertCity(city); // Insere a cidade no banco de dados
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(
-        content: Text("Cidade encontrada!"),
-        duration: Duration(seconds: 1),
-      ),
-    );
-    setState(() {});
+Future<void> _findCity(String city) async {
+  if (await _controller.findCity(city)) {
+    List<City> cities = await _dbService.getAllCities();
+    bool cityExists = cities.any((c) => c.cityName == city && c.historyCities);
+    if (!cityExists) {
+      City cidade = City(cityName: city, historyCities: true);
+      await _dbService.insertCity(cidade);
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(
+          content: Text("Cidade encontrada e adicionada ao histórico!"),
+          duration: Duration(seconds: 1),
+        ),
+      );
+    } else {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(
+          content: Text("Cidade já está no histórico!"),
+          duration: Duration(seconds: 1),
+        ),
+      );
+    }
     Navigator.push(
       context,
-      MaterialPageRoute(builder: (BuildContext context) => DetailsWeatherScreen(city: cityName)),
+      MaterialPageRoute(
+        builder: (BuildContext context) => DetailsWeatherScreen(city: city),
+      ),
     );
   } else {
     ScaffoldMessenger.of(context).showSnackBar(
